+++ include/hall_of_fame.h
@@ -7,7 +7,7 @@ struct HallofFameMon
     u32 personality;
     u16 isShiny:1;
     u16 species:15;
-    u8 lvl;
+    u16 lvl;
     u8 nickname[POKEMON_NAME_LENGTH];
 };
